Replace uri construction by procedure parameter; use https by default
This new parameter accepts a bucket and creates a base URI.  This
allows for more flexible handling of base URIs; for example, Ceph
support two methods of indicating the bucket; either as the first path
component or as the first hostname component.  The former is even the
approach recommended by the Ceph documentation, but it's incompatible
with the original Amazon API!

Anyway, we still use the old "https" parameter in the default
implementation of this new make-base-uri parameter, for backwards
compatibility, but now it defaults to #t.  It is also deprecated: it
will be removed in a future version.